+++ MEP/math.py
@@ -5,36 +5,67 @@ import cmath
 import math
 import random
 from string import ascii_lowercase as letters
-from typing import Any, Callable
+from typing import Any, Callable, TypeVar
 
-from MEP.production import *
+from .production import *
 
+from.formula import Formula, unlock, relock
+
+
+ArgType = TypeVar('ArgType', int, float, complex, bool)
+
+def args_type_check(args):
+    num_count = formula_count = production_count = 0
+    for arg in args:
+        if not isinstance(arg, (Formula, Production)):
+            num_count += 1
+        elif isinstance(arg, Formula):
+            formula_count += 1
+        elif isinstance(arg, Production):
+            production_count += 1
+
+    if num_count == len(args):
+        return 'allnum'
+    elif production_count and formula_count:
+        raise TypeError('unsupported operand type(s) for this math functions: Formula and Production')
+    elif production_count:
+        return 'prod'
+    elif formula_count:
+        return 'form'
 
 def newfunc(func: Callable, name: str):
-    def wrapper(*args: Production | Any):
-        funcs = {}
-        trees = {}
-        args_ = set()
-        unlock(*args)
-        for index, arg in enumerate(args):
-            if isinstance(arg, Production):
-                funcs[index] = arg._func
-                trees[index] = arg._tree
-                args_ = args_ | arg._args
-        relock(*args)
-
-        if not any([isinstance(arg, Production) for arg in args]):
-            return func(*args)
-        return Production(lambda kwargs: func(
-            *[(funcs[index](kwargs) if isinstance(arg, Production) else arg) for index, arg in enumerate(args)]),
-            dict((('S', name),) + tuple(
-                [(letters[index], trees[index] if isinstance(arg, Production) else arg)
-                 for index, arg in enumerate(args)])), 
-            args_)
+    def wrapper(*args: Production | Formula | Any):
+        match args_type_check(args):
+            case 'allnum':
+                return func(*args)
+            case 'prod':
+                funcs, trees, args_ = {}, {}, set()
+                unlock(*args)
+                for index, arg in enumerate(args):
+                    if isinstance(arg, Production):
+                        funcs[index] = arg._func
+                        trees[index] = arg._tree
+                        args_ = args_ | arg._args
+                relock(*args)
+                return Production(lambda kwargs: func(
+                    *[(funcs[index](kwargs) if isinstance(arg, Production) else arg) for index, arg in enumerate(args)]),
+                    dict((('S', name),) + tuple(
+                        [(letters[index], trees[index] if isinstance(arg, Production) else arg)
+                         for index, arg in enumerate(args)])), 
+                    args_)
+            case 'form':
+                productions = {}
+                unlock(*args)
+                for index, arg in enumerate(args):
+                    if isinstance(arg, Formula):
+                        productions[index] = arg._formula._production
+                relock(*args)
+                production = wrapper(*[(productions[index] if isinstance(arg, Formula) else arg) for index, arg in enumerate(args)])
+                return Formula(production)
 
     return wrapper
 
-class Helper:
+class Utils:
 
     @classmethod
     def factorial(cls, x):
@@ -66,7 +97,7 @@ class Helper:
         return random.choices(values, weights)[0]
     
     @classmethod
-    def conditions(cls, *args):
+    def branch(cls, *args):
         if len(args) % 2 != 1:
             raise ValueError('bad value was given')
         for index, item in enumerate(args):
@@ -89,7 +120,7 @@ class Math:
     sqrt = newfunc(cmath.sqrt, 'sqrt')
     cbrt = newfunc(lambda x: x ** (1 / 3), 'cbrt')
     log = newfunc(cmath.log, 'log')
-    fact = newfunc(Helper.factorial, 'fact')
+    fact = newfunc(Utils.factorial, 'fact')
 
     # numtype
     int = newfunc(int, 'int')
@@ -110,7 +141,7 @@ class Math:
     asin = newfunc(cmath.asin, 'asin')
     acos = newfunc(cmath.acos, 'acos')
     atan = newfunc(cmath.atan, 'atan')
-    dist = newfunc(Helper.dist, 'dist')
+    dist = newfunc(Utils.dist, 'dist')
     hypot = newfunc(lambda *args: cmath.sqrt(sum((x ** 2 for x in args))), 'hypot')
 
     # angle
@@ -132,7 +163,7 @@ class Math:
     # random
     rand = newfunc(random.randint, 'rand')
     choose = newfunc(lambda *args: random.choice(args), 'choose')
-    wchoose = newfunc(Helper.wchoose, 'wchoose')
+    wchoose = newfunc(Utils.wchoose, 'wchoose')
 
     # complex
     complex = newfunc(complex, 'complex')
@@ -149,4 +180,4 @@ class Math:
     logicor = newfunc(lambda *args: any(map(bool, args)), 'or')
     logicnot = newfunc(lambda x: not bool(x), 'not')
     logicxor = newfunc(lambda x, y: (not (bool(x) and bool(y))) and (bool(x) or bool(y)), 'xor')
-    conditions = newfunc(Helper.conditions, 'if')
+    branch = newfunc(Utils.branch, 'branch')